Moved driller-worker source from driller/driller -> driller/src
As requested in issue #4
